requirements.txt
fastapi
uvicorn
opencv-python
pillow
numpy
python-multipart
torch
matplotlib
tqdm
albumentations
pydantic